Updates to README, added tweets for paleo and cultural studies, etc (#1)
* Improving README, updating data files

Pulled tweet bodies for mentioned paleo and cultural studies research. New export of tweets for gender studies research. AltmetricID files are lists of the internal identifiers we use at Altmetric; these were parsed from Altmetric Explorer export created on 7 August 2017. Renamed "tweets.csv" and "tweets.json" to be more specific.

* Organized files

* Updates to README docs

Documented what exactly the source data is a bit better.

* Added Journal-level tweet counts data, update to README, comments to analyze.py

* Insignificant changes to analyze.py, remove .DS_Store file

How'd that even get in there...?

* Added stat records for NLTK, etc

* Updated to write tweet text to plaintext

Needed in order to run the same datasets through SentiStrength

* Update to remove incorrect data cleaning documentation

* New SentiStrength input and output files

Note that these should not be the final files; they're based on older tweets and are just meant as a point of reference.